Remove 'FunctionRef' in favor of 'Function'
Function and FunctionRef were unequally applied which raised questions.

It's technically not 100% correct to use 'Function', but 'FunctionRef'
invites more questions that are not answered in this workshop, and not
really appropriate at this level. We should defer that to a more
in-depth workshop where we explain the reasoning behind those.