Update code/ folder to latest Developer Preview
The tableviewer part of the sample is either commented out or removed
for now as that still needs updating. Tested all samples against the
latest release of the AWS CDK Developer Preview and everything comes out
as expected.

This mis-matches the workshop at this point, but at least these will
run/work.