docs: Fix markdown italics (#109)
Converted markdown italics from _underscore_ to the more standard *single-asterisk* format that is used in all other locations in the code.

Also attempted to fix the issue in the 200-lambda.md files where the italics surrounding the "To read more about IDs..." text+link were not being honored when the markdown was being converted to HTML. This resulted in the underscore characters showing on the cdkworkshop.com website and the inner text not being italicized. My fix was to terminate the italicization before the link.