feat(testing): add testing guide to workshop (#131)
This PR adds a new Advanced Topics section to the Typescript portion of the workshop. My idea here was to have an area where we can put topics that aren't necessary for the core workshop, but would still be interesting topics. Let me know if you think there is a better way of organizing this content or if you think it should just be merged into the existing sections.

I've also added an initial chapter on writing unit tests for constructs. The chapter uses the HitCounter construct that is created earlier in the workshop. It walks through writing different types of tests for this construct.