feat: update workshop code based on new init templates (#324)
The bulk of this PR updates the `code` for the workshop based on the newest init templates. Over time little things have changed with the init templates, things like:

- content of `cdk.json`
- content of `pom.xml`
- contents of `tsconfig.json`

The other change is to pin the version of `cdk-dynamo-table-viewer` that we are installing. This is the only 3rd party dependency that we are using in the workshop and we will need to update this package to v2. This change ensures that when we do upgrade that package, the workshop will still install a version that is compatible with v1.