Add notice to compile TypeScript before running test
Add a notice that the updated TypeScript must be compiled again before running the test that validates the change. If the developer has stopped running `npm run watch` when going through this part of the tutorial, they may not know that the TypeScript needs to be recompiled first before the test will pick up the changes made.