Update 200-install.md
Ideally, the package should be installed using requirements.txt so that the requirement will be pushed to the CodeCommit. Then it will know to install it. Current instruction only will install locally, therefore will throw the no module found error will be thrown when we get to the pipeline deploying step. 

But I think the code (line 11-15) need to be reviewed carefully since the preview doesn't show this part correctly. 

{{<highlight python "hl_lines=3">}}
aws-cdk-lib==2.37.0
constructs>=10.0.0,<11.0.0
cdk-dynamo-table-view==0.2.0
{{</highlight>}}