Fix Cookie Banner bug, close #1189
How the cookie banner was presented originally was hacky:

the default style was to use "display:none" and remove the element if it wasn't needed.
This change makes it use a CSS class and an appropriate override.
Functionally, this is identical to the original implementation.

Additionally, the original banner was inside the head of the page.
This caused validation errors and made some browsers upset.
By moving the JS around and making sure it's at the end, the page also loads faster.