Update 200-install.md
The workshop started with `aws-cdk-lib==2.128.0` version then when it came to install `cdk-dynamo-table-view` suddenly the `aws-cdk-lib` upgraded to `2.37.0` which is not compatible with python version 3.10.X. It's better to downgrade it.

Also, you are using `0.2.0` version for `cdk-dynamo-table-view` module which uses` "Runtime": "nodejs12.x"` and nodejs12 is deprecated. I installed `0.2.488` version for `cdk-dynamo-table-view` so now I can deploy the stack seamlessly.

FYI, I'm using PYTHON_3_10 for lambda runtime as the PYTHON_3_7 is also deprecated.

It's better also to upgrade lambda runtime, please.
Could you please update the doc for other fellows too?